requirements.txt
easyocr
opencv-python
streamlit
Pillow
numpy
